# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Mock Interview Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /interviews:
    get:
      operationId: listInterviews
      tags: [interviews]
      summary: List interviews
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed]
      responses:
        "200":
          description: Interview list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Interview"
    post:
      operationId: createInterview
      tags: [interviews]
      summary: Create a new interview session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
...
  /interviews/{id}:
    get:
      operationId: getInterview
      tags: [interviews]
      summary: Get interview details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Interview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
    delete:
      operationId: deleteInterview
      tags: [interviews]
      summary: Delete an interview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /interviews/{id}/start:
    post:
      operationId: startInterview
      tags: [interviews]
      summary: Start an interview session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
  /interviews/{id}/complete:
    post:
      operationId: completeInterview
      tags: [interviews]
      summary: Mark interview as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
  /interviews/{id}/analysis:
    get:
      operationId: getInterviewAnalysis
      tags: [interviews]
      summary: Get AI analysis for a completed interview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewAnalysis"

  # Questions & Answers
  /interviews/{id}/questions:
    get:
      operationId: listInterviewQuestions
      tags: [questions]
      summary: List questions for an interview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Questions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
  /interviews/{id}/answer:
    post:
      operationId: submitAnswer
      tags: [questions]
      summary: Submit an answer to a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerInput"
      responses:
        "200":
          description: Answer evaluated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerResult"
  /interviews/{id}/next-question:
    post:
      operationId: getNextQuestion
      tags: [questions]
      summary: Get next AI-generated question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Next question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"

  # Analytics
...
components:
  schemas:
...
    Interview:
      type: object
      required: [id, userId, title, type, difficulty, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [technical, behavioral, system_design, mixed]
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        status:
          type: string
          enum: [pending, in_progress, completed]
        targetRole:
          type: ["string", "null"]
        overallScore:
          type: ["number", "null"]
        durationMinutes:
          type: ["integer", "null"]
        questionCount:
          type: integer
        answeredCount:
          type: integer
        resumeId:
          type: ["integer", "null"]
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    InterviewInput:
      type: object
      required: [title, type, difficulty]
      properties:
        title:
          type: string
        type:
          type: string
          enum: [technical, behavioral, system_design, mixed]
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        targetRole:
          type: string
        questionCount:
          type: integer
        resumeId:
          type: integer
    InterviewAnalysis:
      type: object
      required: [interviewId, overallScore, summary, strengths, improvements, skillScores]
      properties:
        interviewId:
          type: integer
        overallScore:
          type: number
        summary:
          type: string
        strengths:
          type: array
          items:
            type: string
        improvements:
          type: array
          items:
            type: string
        skillScores:
          type: array
          items:
            $ref: "#/components/schemas/SkillScore"
        detailedFeedback:
          type: ["string", "null"]

    # Questions
    Question:
      type: object
      required: [id, interviewId, text, type, orderIndex]
      properties:
        id:
          type: integer
        interviewId:
          type: integer
        text:
          type: string
        type:
          type: string
          enum: [technical, behavioral, situational, coding]
        orderIndex:
          type: integer
        hint:
          type: ["string", "null"]
        userAnswer:
          type: ["string", "null"]
        score:
          type: ["number", "null"]
        feedback:
          type: ["string", "null"]
        isAnswered:
          type: boolean
    AnswerInput:
      type: object
      required: [questionId, answer]
      properties:
        questionId:
          type: integer
        answer:
          type: string
    AnswerResult:
      type: object
      required: [questionId, score, feedback]
      properties:
        questionId:
          type: integer
        score:
          type: number
        feedback:
          type: string
        strengths:
          type: array
          items:
            type: string
        improvements:
          type: array
          items:
            type: string

    # Analytics
...
    SkillScore:
      type: object
      required: [skill, score, maxScore]
      properties:
        skill:
          type: string
        score:
          type: number
        maxScore:
          type: number
...