# lib/api-zod/src/generated/types/answerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Mock Interview Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AnswerInput {
  questionId: number;
  answer: string;
}